# backend/openapi.yaml
openapi: "3.1.0"
info:
  title: SolarGrid Backend API
  description: |
    REST API for the Stellar SolarGrid PAYG solar energy platform.

    Provides access to smart meter data, payment processing, energy statistics,
    provider revenue, webhook management, and admin utilities.  Most write
    endpoints and admin endpoints require the `X-Admin-Key` header.
  version: "0.1.0"
  license:
    name: MIT
servers:
  - url: http://localhost:3001
    description: Local development server

# ── Security ──────────────────────────────────────────────────────────────────
components:
  securitySchemes:
    AdminKey:
      type: apiKey
      in: header
      name: X-Admin-Key
      description: >
        Admin API key set via the `ADMIN_API_KEY` environment variable.
        Required for all write and admin-only endpoints.
  schemas:
...
    ContractStats:
      type: object
      required: [totalMeters, activeMeters, totalUnits, totalRevenue]
      properties:
        totalMeters:
          type: integer
          description: Total registered meters on-chain
        activeMeters:
          type: integer
          description: Meters currently active (paid and within validity period)
        totalUnits:
          type: integer
          description: Total energy units consumed across all meters
        totalRevenue:
          type: integer
          description: Provider total revenue in stroops
    MetricsSummary:
      type: object
      required: [mqttMessages, contractCalls, activeMeters, paymentVolumeXlm]
      properties:
        mqttMessages:
          type: number
          description: Total MQTT messages received since process start
        contractCalls:
          type: number
          description: Total Soroban contract invocations since process start
        activeMeters:
          type: number
          description: Current active meter gauge value
        paymentVolumeXlm:
          type: number
          description: Cumulative payment volume in XLM

    # ── Provider ──────────────────────────────────────────────────────────────
...
paths:
...
  /api/stats:
    get:
      summary: Contract-derived meter statistics
      description: |
        Queries the Soroban contract for a real-time aggregate of all meters:
        total count, active count, total units consumed, and provider revenue.

        Response is cached for **30 seconds** to reduce RPC load.
      operationId: getStats
      responses:
        "200":
          description: Meter statistics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContractStats"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/stats/summary:
    get:
      summary: Prom-client metrics snapshot
      description: |
        Returns a JSON snapshot of the four core prom-client counters/gauges
        for the admin dashboard.  Does **not** require Prometheus or Grafana
        to be running.

        Response is cached for **15 seconds**.
      operationId: getStatsSummary
      responses:
        "200":
          description: Metrics snapshot
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MetricsSummary"
        "500":
          $ref: "#/components/responses/InternalError"

  # ── Provider ─────────────────────────────────────────────────────────────────
...
  /api/provider/webhook:
    post:
      summary: Register provider low-balance webhook
      description: |
        Register or replace the webhook URL that receives low-balance
        notifications when a meter's balance drops below `LOW_BALANCE_THRESHOLD`.

        Alias for `POST /api/webhooks/low-balance` in the provider-facing
        namespace.  Requires admin authentication.
      operationId: registerProviderWebhook
      security:
        - AdminKey: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [webhook_url]
              properties:
                webhook_url:
                  type: string
                  format: uri
                  example: "https://provider.example.com/hooks/low-balance"
      responses:
        "200":
          description: Webhook registered
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  webhook_url:
                    type: string
                    format: uri
        "400":
          $ref: "#/components/responses/ValidationError"
        "401":
          $ref: "#/components/responses/Unauthorized"

  # ── Meters ────────────────────────────────────────────────────────────────────
...
  /api/webhooks/low-balance:
    post:
      summary: Register low-balance webhook URL
      description: |
        Register a webhook URL to receive `low_balance` notifications.
        Multiple URLs may be registered; all receive the notification.
      operationId: registerLowBalanceWebhook
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [webhook_url]
              properties:
                webhook_url:
                  type: string
                  format: uri
      responses:
        "200":
          description: Webhook registered
        "400":
          $ref: "#/components/responses/ValidationError"
  /api/webhooks/sms-payment:
    post:
      summary: Telecom partner SMS payment webhook
      description: |
        Receives a signed payment event from a telecom partner and triggers
        `make_payment` on-chain.  Requires a valid `X-Webhook-Signature` header
        (HMAC-SHA256 of the raw request body using `TELECOM_WEBHOOK_SECRET`).
      operationId: handleSmsPaymentWebhook
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [meter_id, amount_xlm]
              properties:
                meter_id: { type: string }
                amount_xlm: { type: number }
                plan:
                  type: string
                  enum: [Daily, Weekly, Usage, UsageBased]
                  default: Daily
      responses:
        "200":
          $ref: "#/components/responses/TxHash"
        "401":
          description: Invalid webhook signature

  # ── Allowlist ─────────────────────────────────────────────────────────────────
...